Cells v1.1 -- Initial commit
A real-time synthtizer composed from a matrix of interacting cells.
The cells themselves are composed of a modulated oscillator
and an effects chain, both of which are randomly generated
from a built-in library of specific behaviours.
The oscillator's modulation inputs are taken from the outputs
of the neighbouring cells, in random order, using a feedback loop.
The matrix's output is routed to SC's output through the selected
output module. Different output modules are provided.